pyproject.toml
[project]
name = "spotlightserver"
version = "0.1.0"
description = "API server for Spotlight Mobile Application"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic==0.7.7",
    "fastapi==0.104.1",
    "google-generativeai==0.3.1",
    "gunicorn==21.2.0",
    "httpx==0.25.2",
    "openai==1.3.7",
    "pydantic==2.5.0",
    "python-dotenv==1.0.0",
    "uvicorn[standard]==0.24.0",
]
